requirements.txt
fastapi
uvicorn
langchain==0.1.16
langchain-community==0.0.32
langchain-core
langchain-google-genai
faiss-cpu==1.8.0
sentence-transformers
pydantic==2.7.1
python-dotenv==1.0.0
google-generativeai==0.5.0